refactor(core): split the sandboxed-exec engine into a codex-sandbox crate (#527)
Dissolve the codex-core god crate into two honest crates:

- codex-sandbox (new, 3.1k LOC): the pure exec engine —
  process_exec_tool_call, spawn, platform sandbox selection
  (Seatbelt/Landlock/Windows), env policy application, output encoding and
  truncation, and the CodexErr/SandboxErr types. No config or auth
  dependencies; the policy types it consumes (the ShellEnvironmentPolicy
  family) are hoisted to codex_protocol::config_types. Its
  exec/seatbelt/text_encoding integration tests move with it.
- codex-core (11.3k LOC remaining): config and config_loader,
  auth/token_data, MCP auth helpers, git_info, model metadata, project_doc,
  features, and otel_init. It depends on codex-sandbox, matching the
  pre-existing call direction.

Also deleted: TruncationPolicy::new(&Config), four uncalled truncate helpers,
and TruncationPolicy::mul (zero callers). Renaming codex-core itself is
deferred to a dedicated pass.

Building codex-sandbox standalone caught two missing tokio features
previously masked by workspace feature unification — the independent-crate
payoff working as intended.
