+++ nori-rs/core/Cargo.toml
@@ -14,10 +14,8 @@ workspace = true
 
 [dependencies]
 anyhow = { workspace = true }
-async-channel = { workspace = true }
 base64 = { workspace = true }
 chrono = { workspace = true, features = ["serde"] }
-chardetng = { workspace = true }
 codex-app-server-protocol = { workspace = true }
 codex-apply-patch = { workspace = true }
 codex-async-utils = { workspace = true }
@@ -27,20 +25,18 @@ codex-git = { workspace = true }
 codex-keyring-store = { workspace = true }
 codex-otel = { workspace = true, features = ["otel"] }
 codex-protocol = { workspace = true }
+codex-sandbox = { workspace = true }
 codex-rmcp-client = { workspace = true }
 codex-utils-pty = { workspace = true }
 codex-utils-readiness = { workspace = true }
 codex-utils-string = { workspace = true }
-codex-windows-sandbox = { package = "codex-windows-sandbox", path = "../windows-sandbox-rs" }
 diffy = { workspace = true }
 dirs = { workspace = true }
 dunce = { workspace = true }
 env-flags = { workspace = true }
-encoding_rs = { workspace = true }
 futures = { workspace = true }
 http = { workspace = true }
 keyring = { workspace = true, features = ["crypto-rust"] }
-libc = { workspace = true }
 mcp-types = { workspace = true }
 os_info = { workspace = true }
 rand = { workspace = true }
@@ -71,7 +67,6 @@ tokio = { workspace = true, features = [
     "rt-multi-thread",
     "signal",
 ] }
-tokio-util = { workspace = true, features = ["rt"] }
 toml = { workspace = true }
 toml_edit = { workspace = true }
 tracing = { workspace = true, features = ["log"] }
@@ -86,8 +81,6 @@ deterministic_process_ids = []
 
 
 [target.'cfg(target_os = "linux")'.dependencies]
-landlock = { workspace = true }
-seccompiler = { workspace = true }
 keyring = { workspace = true, features = ["linux-native-async-persistent"] }
 
 [target.'cfg(target_os = "macos")'.dependencies]
@@ -124,7 +117,6 @@ codex-core = { path = ".", features = ["deterministic_process_ids"] }
 core_test_support = { workspace = true }
 ctor = { workspace = true }
 escargot = { workspace = true }
-maplit = { workspace = true }
 predicates = { workspace = true }
 pretty_assertions = { workspace = true }
 serial_test = { workspace = true }
