+++ nori-rs/core/tests/suite/auth_refresh.rs
@@ -10,9 +10,9 @@ use codex_core::auth::REFRESH_TOKEN_URL_OVERRIDE_ENV_VAR;
 use codex_core::auth::RefreshTokenError;
 use codex_core::auth::load_auth_dot_json;
 use codex_core::auth::save_auth;
-use codex_core::error::RefreshTokenFailedReason;
 use codex_core::token_data::IdTokenInfo;
 use codex_core::token_data::TokenData;
+use codex_sandbox::error::RefreshTokenFailedReason;
 use core_test_support::skip_if_no_network;
 use pretty_assertions::assert_eq;
 use serde::Serialize;
