+++ nori-rs/docs.md
@@ -12,14 +12,14 @@ The `nori-rs` directory is the root of a Cargo workspace containing all Rust cod
 
 - **Entry points**: `tui/` provides the main TUI application, `cli/` provides the `nori` binary dispatch and sandbox debug utilities
 - **ACP integration**: `acp/` handles communication with ACP-compliant agents spawned as local subprocesses, and owns session state, transcripts, Nori config, and the moved-in leaf helpers (shell parsing, notifications, custom prompts, compact constants)
-- **Shared infrastructure**: `core/` contains configuration, authentication, and sandboxed command execution consumed by the frontends (not by `acp/`)
+- **Shared infrastructure**: `core/` contains configuration, authentication, and model/provider metadata consumed by the frontends (not by `acp/`)
 - **Protocol definitions**: `protocol/`, `app-server-protocol/`, `mcp-types/` define shared type vocabularies
-- **Sandboxing**: `linux-sandbox/`, `execpolicy/` provide command execution security
+- **Sandboxing**: `sandbox/` (`codex-sandbox`) owns the sandboxed exec engine and platform sandbox selection; `linux-sandbox/`, `windows-sandbox-rs/`, `execpolicy/` provide the platform-specific pieces
 - **Utilities**: Various crates in `utils/` provide shared functionality
 
 Most shared crates still follow the inherited `codex-` prefix convention (for example `codex-core` and `codex-protocol`), while Nori-owned entrypoint crates now use `nori-` names such as `nori-acp`, `nori-protocol`, `nori-installed`, and `nori-tui`.
 
-The workspace is converging on a three-layer structure -- publishable ACP-host leaves at the bottom, a headless session harness in the middle, thin frontends on top -- per `@/docs/specs/crate-layering.md`. Milestones already landed: dead Codex-engine subsystems deleted from `codex-core`, the `nori-tui` `nori-config` cargo feature removed (Nori config is the only path), protocol types imported directly from `codex-protocol` (core's re-export detour deleted), and `nori-acp`'s dependency on `codex-core` fully severed.
+The workspace is converging on a three-layer structure -- publishable ACP-host leaves at the bottom, a headless session harness in the middle, thin frontends on top -- per `@/docs/specs/crate-layering.md`. Milestones already landed: dead Codex-engine subsystems deleted from `codex-core`, the `nori-tui` `nori-config` cargo feature removed (Nori config is the only path), protocol types imported directly from `codex-protocol` (core's re-export detour deleted), `nori-acp`'s dependency on `codex-core` fully severed, and the sandboxed-execution engine extracted from `codex-core` into `codex-sandbox` (`@/nori-rs/sandbox/`).
 
 ### Core Implementation
 
@@ -28,7 +28,8 @@ The TUI drives user interaction through a Ratatui-based interface. When using AC
 Architecture:
 - nori-tui (TUI) -> Terminal User Interface
   - nori-acp -> ACP Agent Connection -> External ACP Agents (claude, etc); depends only on codex-protocol among inherited crates
-  - codex-core -> Config/Auth/Sandboxing infrastructure for the frontends (nori-tui, nori-cli)
+  - codex-core -> Config/Auth infrastructure for the frontends (nori-tui, nori-cli)
+  - codex-sandbox -> Sandboxed exec engine and platform sandbox selection; imported directly by core, tui, cli, and linux-sandbox
   - codex-protocol -> Shared type vocabulary, imported directly by every consumer
 
 ### Things to Know
