+++ nori-rs/linux-sandbox/Cargo.toml
@@ -18,6 +18,7 @@ workspace = true
 [target.'cfg(target_os = "linux")'.dependencies]
 clap = { workspace = true, features = ["derive"] }
 codex-core = { workspace = true }
+codex-sandbox = { workspace = true }
 codex-protocol = { workspace = true }
 landlock = { workspace = true }
 libc = { workspace = true }
