+++ nori-rs/linux-sandbox/src/landlock.rs
@@ -2,10 +2,10 @@ use std::collections::BTreeMap;
 use std::path::Path;
 use std::path::PathBuf;
 
-use codex_core::error::CodexErr;
-use codex_core::error::Result;
-use codex_core::error::SandboxErr;
 use codex_protocol::protocol::SandboxPolicy;
+use codex_sandbox::error::CodexErr;
+use codex_sandbox::error::Result;
+use codex_sandbox::error::SandboxErr;
 
 use landlock::ABI;
 use landlock::Access;
