+++ nori-rs/linux-sandbox/tests/suite/landlock.rs
@@ -1,11 +1,11 @@
 #![cfg(target_os = "linux")]
 use codex_core::config::types::ShellEnvironmentPolicy;
-use codex_core::error::CodexErr;
-use codex_core::error::SandboxErr;
-use codex_core::exec::ExecParams;
-use codex_core::exec::process_exec_tool_call;
-use codex_core::exec_env::create_env;
 use codex_protocol::protocol::SandboxPolicy;
+use codex_sandbox::error::CodexErr;
+use codex_sandbox::error::SandboxErr;
+use codex_sandbox::exec::ExecParams;
+use codex_sandbox::exec::process_exec_tool_call;
+use codex_sandbox::exec_env::create_env;
 use std::collections::HashMap;
 use std::path::PathBuf;
 use tempfile::NamedTempFile;
