+++ nori-rs/protocol/Cargo.toml
@@ -12,6 +12,7 @@ path = "src/lib.rs"
 workspace = true
 
 [dependencies]
+wildmatch = { workspace = true }
 codex-git = { workspace = true }
 
 base64 = { workspace = true }
