+++ nori-rs/sandbox/tests/exec.rs
@@ -3,17 +3,17 @@
 use std::collections::HashMap;
 use std::string::ToString;
 
-use codex_core::exec::ExecParams;
-use codex_core::exec::ExecToolCallOutput;
-use codex_core::exec::SandboxType;
-use codex_core::exec::process_exec_tool_call;
-use codex_core::spawn::CODEX_SANDBOX_ENV_VAR;
 use codex_protocol::protocol::SandboxPolicy;
+use codex_sandbox::exec::ExecParams;
+use codex_sandbox::exec::ExecToolCallOutput;
+use codex_sandbox::exec::SandboxType;
+use codex_sandbox::exec::process_exec_tool_call;
+use codex_sandbox::spawn::CODEX_SANDBOX_ENV_VAR;
 use tempfile::TempDir;
 
-use codex_core::error::Result;
+use codex_sandbox::error::Result;
 
-use codex_core::get_platform_sandbox;
+use codex_sandbox::get_platform_sandbox;
 
 fn skip_test() -> bool {
     if std::env::var(CODEX_SANDBOX_ENV_VAR) == Ok("seatbelt".to_string()) {
